requirements.txt
djangorestframework
drf-yasg
python-decouple
mysqlclient
django-cors-headers